Fix compilation errors in FamilyUserDeviceMetricsTest.
After https://crrev.com/c/2676393 was landed, I receive the following
error when building browser_tests locally:
../../chrome/browser/chromeos/child_accounts/family_user_device_metrics_browsertest.cc:199:3: note: in instantiation of function template specialization 'testing::internal::EqHelper::Compare<int, unsigned long, nullptr>' requested here
  EXPECT_EQ(total_user_count, user_manager_->GetUsers().size());

This is caused by some extra compiler configuration that is not set
on regular trybots, but may be set on others.

R=agawronska@chromium.org, xiqiruan@chromium.org

Bug: 1172335
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2713806
Auto-Submit: Sergey Poromov <poromov@chromium.org>
Commit-Queue: Aga Wronska <agawronska@chromium.org>
Cr-Commit-Position: refs/heads/master@{#856713}

+++ chrome/browser/chromeos/child_accounts/family_user_device_metrics_browsertest.cc
@@ -196,7 +196,7 @@ IN_PROC_BROWSER_TEST_P(FamilyUserDeviceMetricsTest, GuestUser) {
   logged_in_user_mixin_.GetLoginManagerMixin()->LoginAsNewRegularUser();
   logged_in_user_mixin_.GetLoginManagerMixin()->WaitForActiveSession();
 
-  int total_user_count = IsUserExisting() ? 3 : 2;
+  size_t total_user_count = IsUserExisting() ? 3 : 2;
   EXPECT_EQ(total_user_count, user_manager_->GetUsers().size());
 
   // If no existing users on login screen, then this user is the first and only.
@@ -215,7 +215,7 @@ IN_PROC_BROWSER_TEST_P(FamilyUserDeviceMetricsTest, ActiveDirectoryUser) {
   logged_in_user_mixin_.GetLoginManagerMixin()->LoginAsNewRegularUser();
   logged_in_user_mixin_.GetLoginManagerMixin()->WaitForActiveSession();
 
-  int total_user_count = IsUserExisting() ? 3 : 2;
+  size_t total_user_count = IsUserExisting() ? 3 : 2;
   EXPECT_EQ(total_user_count, user_manager_->GetUsers().size());
 
   // If no existing users on login screen, then this user is the first and only.
@@ -252,7 +252,7 @@ IN_PROC_BROWSER_TEST_P(FamilyUserDeviceMetricsManagedDeviceTest, KioskAppUser) {
   user_manager_->AddKioskAppUser(kManagedUserAccountId);
   LoginAsNewRegularUser();
 
-  int total_user_count = IsUserExisting() ? 3 : 2;
+  size_t total_user_count = IsUserExisting() ? 3 : 2;
   EXPECT_EQ(total_user_count, user_manager_->GetUsers().size());
 
   // If no existing users on login screen, then this user is the first and only.
@@ -271,7 +271,7 @@ IN_PROC_BROWSER_TEST_P(FamilyUserDeviceMetricsManagedDeviceTest,
   user_manager_->AddArcKioskAppUser(kManagedUserAccountId);
   LoginAsNewRegularUser();
 
-  int total_user_count = IsUserExisting() ? 3 : 2;
+  size_t total_user_count = IsUserExisting() ? 3 : 2;
   EXPECT_EQ(total_user_count, user_manager_->GetUsers().size());
 
   // If no existing users on login screen, then this user is the first and only.
@@ -290,7 +290,7 @@ IN_PROC_BROWSER_TEST_P(FamilyUserDeviceMetricsManagedDeviceTest,
   user_manager_->AddWebKioskAppUser(kManagedUserAccountId);
   LoginAsNewRegularUser();
 
-  int total_user_count = IsUserExisting() ? 3 : 2;
+  size_t total_user_count = IsUserExisting() ? 3 : 2;
   EXPECT_EQ(total_user_count, user_manager_->GetUsers().size());
 
   // If no existing users on login screen, then this user is the first and only.
@@ -309,7 +309,7 @@ IN_PROC_BROWSER_TEST_P(FamilyUserDeviceMetricsManagedDeviceTest,
   user_manager_->AddPublicAccountUser(kManagedUserAccountId);
   LoginAsNewRegularUser();
 
-  int total_user_count = IsUserExisting() ? 3 : 2;
+  size_t total_user_count = IsUserExisting() ? 3 : 2;
   EXPECT_EQ(total_user_count, user_manager_->GetUsers().size());
 
   // If no existing users on login screen, then this user is the first and only.
